requirements.txt
python-dotenv
pandas
numpy
pymongo
certifi
pymongo[srv]
scikit-learn
PyYAML
mlflow
dagshub
FastAPI
uvicorn
pytest
pytest-cov
jinja2
python-multipart
dill
scipy

## -e .